issue/1469 Variable names, filename and imports
