+++ js/TUTOR_TYPE.js
